fix(decomp): Spell three definitions' parameter types as their declarations do
xModelGetBoneLocation took size_t where xModel.h says u32, and two xDynAsset
loaders took unsigned long where their declarations say size_t. Same types
on GameCube; they differ on other ABIs.

+++ src/SB/Core/x/xEntBoulder.cpp
@@ -1077,7 +1077,7 @@ static void RecurseChild(xBase* child, xEntBoulder** boulList, S32& currBoul)
     }
 }
 
-void xBoulderGenerator_Init(xBase& data, xDynAsset& asset, unsigned long)
+void xBoulderGenerator_Init(xBase& data, xDynAsset& asset, size_t)
 {
     xBoulderGenerator_Init((xBoulderGenerator*)&data, (xBoulderGeneratorAsset*)&asset);
 }
